Move write_uvdata function to utils
