+++ pyuvsim/tests/test_utils.py
@@ -6,17 +6,18 @@ from __future__ import absolute_import, division, print_function
 
 import numpy as np
 import os
+import shutil
 import nose.tools as nt
 from astropy.time import Time
 from astropy.coordinates import Angle
 
+from pyuvdata import UVData
+import pyuvdata.tests as uvtest
+
+from pyuvsim.data import DATA_PATH as SIM_DATA_PATH
 from pyuvsim import utils as simutils
 
-import pyuvsim
-from pyuvsim.data import DATA_PATH as SIM_DATA_PATH
-
-# Five different test configs
-param_filenames = [os.path.join(SIM_DATA_PATH, 'test_config', 'param_10time_10chan_{}.yaml'.format(x)) for x in range(5)]
+triangle_uvfits_file = os.path.join(SIM_DATA_PATH, '28m_triangle_10time_10chan.uvfits')
 
 
 def test_tee_ra_loop():
@@ -106,8 +107,32 @@ def test_file_namer():
             f.write(' ')
         fnames.append(fname)
     existing_file = fnames[0]
-    new_filepath = pyuvsim.utils.check_file_exists_and_increment(existing_file)
+    new_filepath = simutils.check_file_exists_and_increment(existing_file)
     for fn in fnames:
         os.remove(fn)
     os.rmdir('tempfiles')
     nt.assert_true(new_filepath.endswith("_111"))    # There are four other of these param test files
+
+
+def test_write_uvdata():
+    """ Test function that defines filenames from parameter dict """
+
+    uv = UVData()
+    uvtest.checkWarnings(uv.read_uvfits, [triangle_uvfits_file],
+                         message='Telescope 28m_triangle_10time_10chan.yaml is not in known_telescopes.')
+
+    ofname = 'test_file'
+    filing_dict = {'outfile_name': ofname}
+    expected_ofname = simutils.write_uvdata(uv, filing_dict, return_filename=True)
+    ofname = os.path.join('.', ofname)
+    nt.assert_equal(ofname + '.uvfits', expected_ofname)
+    expected_ofname = simutils.write_uvdata(uv, filing_dict, return_filename=True, out_format='miriad')
+    nt.assert_equal(ofname, expected_ofname)
+    nt.assert_raises(ValueError, simutils.write_uvdata, uv, filing_dict, return_filename=True, out_format='')
+    filing_dict['output_format'] = 'uvh5'
+    expected_ofname = simutils.write_uvdata(uv, filing_dict, return_filename=True)
+    nt.assert_equal(ofname + '.uvh5', expected_ofname)
+
+    os.remove(ofname + '.uvh5')
+    os.remove(ofname + '.uvfits')
+    shutil.rmtree(ofname)
